split out stuff into its relevant testclass
