+++ test/test_helper.rb
@@ -9,17 +9,17 @@ require File.expand_path('../../config/environment', __FILE__)
 require 'rails/test_help'
 require 'minitest/reporters'
 require 'minitest/autorun'
-require 'capybara/rails'
+# require 'capybara/rails'
 require 'capybara/poltergeist'
-require 'capybara-screenshot/minitest'
+# require 'capybara-screenshot/minitest'
 require 'omniauth_helper'
 require 'rack/test'
 require 'integration_system_test_helpers'
 Minitest::Reporters.use!
 
-Capybara.register_driver :poltergeist do |app|
-  Capybara::Poltergeist::Driver.new(app, js_errors: false)
-end
+# Capybara.register_driver :poltergeist do |app|
+  # Capybara::Poltergeist::Driver.new(app, js_errors: false)
+# end
 
 DatabaseCleaner.clean_with :truncation
 
@@ -34,9 +34,17 @@ end
 Capybara.save_path = "#{ENV.fetch('CIRCLE_ARTIFACTS', Rails.root.join('tmp/capybara'))}" if ENV['CIRCLE_ARTIFACTS']
 
 class ActionDispatch::IntegrationTest
-  include IntegrationSystemTestHelpers
+  # include IntegrationSystemTestHelpers
 
-  before { Capybara.reset_sessions! }
+  def sign_in(user)
+    post user_session_path \
+      'user[email]' => user.email,
+      'user[password]' => user.password
+  end
+
+  def choose_line(line)
+    post lines_path, params: { line: line.to_s }
+  end
 end
 
 class ActionController::TestCase
